pyproject.toml
[project]
name = "pipegate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "async-timeout>=5.0.1",
    "fastapi[standard]>=0.115.6",
    "httpx>=0.28.1",
    "orjson>=3.10.12",
    "pydantic-settings>=2.7.0",
    "pydantic>=2.10.3",
    "pyjwt>=2.10.1",
    "typer>=0.15.1",
    "uvicorn>=0.32.1",
    "websockets>=14.1",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "ruff>=0.8.2",
]

[tool.mypy]
plugins = ['pydantic.mypy']